Add files via upload
The idea behind this project is to implement a backend data generating script and display this generated data on the front end webpage. This method should be useful in upcoming projects where data is searched and access by the user through using a simple search box functionality on the website.